Validateの厳密化及び厳密化のmodel_spec作成

+++ db/migrate/20210329064040_create_products.rb
@@ -8,7 +8,7 @@ class CreateProducts < ActiveRecord::Migration[5.2]
       t.integer :selling_price, null: false
       t.integer :size, null: false
       t.integer :storage_size, null: false
-      t.integer :weight, null: false
+      t.float :weight, null: false
       t.integer :stock, default: 1
       t.string :brand
